requirements.txt
asgiref==3.4.1
autopep8==1.6.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
Django==4.0
django-cors-headers==3.11.0
django-filter==21.1
django-rest-knox==4.1.0
djangorestframework==3.13.1
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.14.1
mysqlclient==2.1.0
Pillow==9.0.0
pycodestyle==2.8.0
pycparser==2.21
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.29
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
Werkzeug==2.0.2